Force data into an object; work with object keys/values throughout
Previously we would work with the data exclusively as an array. If an
object was passed in for the data, we would coerce it into an
array. This update reverses that; if we get an array, we coerce it
into an object, making up property names/labels for it. Then we work
with the object throughout the code. This is to prepare for dealing
with stacked chart data. The code for traversing an object is a lot
cleaner and easier to understand than the code for traversing a
multi-dimensional array.